Cast the parameters of boost::endian::endian_reverse(), just in case.

+++ taglib/toolkit/tutils.h
@@ -65,7 +65,7 @@ namespace TagLib
       {
 #if defined(HAVE_BOOST_BYTESWAP)
 
-        return boost::endian::endian_reverse(x);
+        return boost::endian::endian_reverse(static_cast<uint16_t>(x));
 
 #elif defined(HAVE_GCC_BYTESWAP)
 
@@ -101,7 +101,7 @@ namespace TagLib
       {
 #if defined(HAVE_BOOST_BYTESWAP)
 
-        return boost::endian::endian_reverse(x);
+        return boost::endian::endian_reverse(static_cast<uint32_t>(x));
 
 #elif defined(HAVE_GCC_BYTESWAP)
 
@@ -140,7 +140,7 @@ namespace TagLib
       {
 #if defined(HAVE_BOOST_BYTESWAP)
 
-        return boost::endian::endian_reverse(x);
+        return boost::endian::endian_reverse(static_cast<uint64_t>(x));
 
 #elif defined(HAVE_GCC_BYTESWAP)
